+++ cc/titers.hh
@@ -181,7 +181,7 @@ namespace acmacs::chart
         virtual size_t number_of_antigens() const = 0;
         virtual size_t number_of_sera() const = 0;
         virtual size_t number_of_non_dont_cares() const = 0;
-        virtual double percent_of_non_dont_cares() const { return static_cast<double>(number_of_non_dont_cares()) / (number_of_antigens() * number_of_sera()); }
+        virtual double percent_of_non_dont_cares() const { return static_cast<double>(number_of_non_dont_cares()) / static_cast<double>(number_of_antigens() * number_of_sera()); }
         virtual bool is_dense() const noexcept;
 
           // support for fast exporting into ace, if source was ace or acd1
